requirements.txt
customtkinter
cryptography
bcrypt
pyotp
qrcode
Pillow
python-dotenv
Flask   
Flask-Cors
pyinstaller